+++ src/Consts.ts
@@ -1,5 +1,7 @@
-export const LEGACY_VSCODE_CONFIG_MESSAGE =
+export const LEGACY_VSCODE_PRETTIER_CONFIG_MESSAGE =
   "You have legacy settings in your VS Code config. They are being ignored Would you like to migrate them to '.prettierrc'?.";
+export const LEGACY_VSCODE_LINTER_CONFIG_MESSAGE =
+  "You have legacy linter settings in your VS Code config. They are no longer being used. See documentation for migration information.";
 export const OUTDATED_PRETTIER_VERSION_MESSAGE =
   "Outdated version of prettier detected. Upgrade to latest.";
 export const MIGRATE_CONFIG_ACTION_TEXT = "Migrate Config";