+++ src/NotificationService.ts
@@ -1,9 +1,16 @@
-// tslint:disable-next-line: no-implicit-dependencies
-import { ConfigurationTarget, Uri, window, workspace } from "vscode";
+import {
+  ConfigurationTarget,
+  Uri,
+  window,
+  workspace,
+  WorkspaceConfiguration
+  // tslint:disable-next-line: no-implicit-dependencies
+} from "vscode";
 import * as nls from "vscode-nls";
 import { createConfigFileFunction } from "./Commands";
 import {
-  LEGACY_VSCODE_CONFIG_MESSAGE,
+  LEGACY_VSCODE_LINTER_CONFIG_MESSAGE,
+  LEGACY_VSCODE_PRETTIER_CONFIG_MESSAGE,
   MIGRATE_CONFIG_ACTION_TEXT,
   OUTDATED_PRETTIER_VERSION_MESSAGE
 } from "./Consts";
@@ -32,57 +39,13 @@ export class NotificationService {
     if (this.noLegacyConfigWorkspaces.indexOf(cacheKey) > -1) {
       return;
     }
-
     const vscodeConfig = workspace.getConfiguration("prettier", uri);
-    const legacyConfigOptions = new Map<string, any>();
-    legacyConfigOptions.set("printWidth", 80);
-    legacyConfigOptions.set("tabWidth", 2);
-    legacyConfigOptions.set("singleQuote", false);
-    legacyConfigOptions.set("trailingComma", "none");
-    legacyConfigOptions.set("bracketSpacing", true);
-    legacyConfigOptions.set("jsxBracketSameLine", false);
-    legacyConfigOptions.set("semi", true);
-    legacyConfigOptions.set("useTabs", false);
-    legacyConfigOptions.set("proseWrap", "preserve");
-    legacyConfigOptions.set("arrowParens", "avoid");
-    legacyConfigOptions.set("jsxSingleQuote", false);
-    legacyConfigOptions.set("htmlWhitespaceSensitivity", "css");
-    legacyConfigOptions.set("endOfLine", "auto");
-    legacyConfigOptions.set("quoteProps", "as-needed");
-
-    const migratedOptions = new Map<string, any>();
-    legacyConfigOptions.forEach((defaultValue, key) => {
-      const val = vscodeConfig.get(key);
-      if (val !== defaultValue) {
-        migratedOptions.set(key, val);
-      }
-    });
-    if (migratedOptions.size === 0) {
-      // No legacy configs, add to
+    const hasLegacyConfig =
+      (await this.warnIfLegacyPrettierConfiguration(vscodeConfig)) ||
+      (await this.warnIfLegacyLinterConfiguration(vscodeConfig));
+    if (!hasLegacyConfig) {
+      // No legacy configs, add to cache
       this.noLegacyConfigWorkspaces.push(cacheKey);
-    } else {
-      const message = localize(
-        "ext.config.legacyVSCodeConfigInUse",
-        LEGACY_VSCODE_CONFIG_MESSAGE
-      );
-
-      const result = await window.showErrorMessage(
-        message,
-        MIGRATE_CONFIG_ACTION_TEXT
-      );
-      if (result && result === MIGRATE_CONFIG_ACTION_TEXT) {
-        await this.createConfigFileCommand(migratedOptions);
-        legacyConfigOptions.forEach((val, key) => {
-          const inspected = vscodeConfig.inspect(key);
-          if (inspected && inspected.workspaceFolderValue) {
-            vscodeConfig.update(
-              key,
-              undefined,
-              ConfigurationTarget.WorkspaceFolder
-            );
-          }
-        });
-      }
     }
   }
 
@@ -100,5 +63,90 @@ export class NotificationService {
     } else {
       return window.showErrorMessage(localizedMessage);
     }
+  }
+
+  public clearCache() {
+    this.noLegacyConfigWorkspaces = [];
+  }
+
+  /**
+   * Check if the resource has legacy config. Returns true if legacy config is found.
+   * @param vscodeConfig The configuration
+   */
+  private async warnIfLegacyLinterConfiguration(
+    vscodeConfig: WorkspaceConfiguration
+  ): Promise<boolean> {
+    const hasLegacyConfig =
+      vscodeConfig.get("eslintIntegration") !== null ||
+      vscodeConfig.get("tslintIntegration") !== null ||
+      vscodeConfig.get("stylelintIntegration") !== null;
+    if (hasLegacyConfig) {
+      const message = localize(
+        "ext.config.legacyLinterConfigInUse",
+        LEGACY_VSCODE_LINTER_CONFIG_MESSAGE
+      );
+
+      await window.showErrorMessage(message);
+    }
+    return hasLegacyConfig;
+  }
+
+  /**
+   * Check if the resource has legacy config. Returns true if legacy config is found.
+   * @param vscodeConfig The configuration
+   */
+  private async warnIfLegacyPrettierConfiguration(
+    vscodeConfig: WorkspaceConfiguration
+  ): Promise<boolean> {
+    const legacyConfigOptions = [
+      "printWidth",
+      "tabWidth",
+      "singleQuote",
+      "trailingComma",
+      "bracketSpacing",
+      "jsxBracketSameLine",
+      "semi",
+      "useTabs",
+      "proseWrap",
+      "arrowParens",
+      "jsxSingleQuote",
+      "htmlWhitespaceSensitivity",
+      "endOfLine",
+      "quoteProps"
+    ];
+
+    const migratedOptions = new Map<string, any>();
+    legacyConfigOptions.forEach(key => {
+      const val = vscodeConfig.get(key);
+      if (val !== null) {
+        migratedOptions.set(key, val);
+      }
+    });
+    const hasLegacyConfig = migratedOptions.size > 0;
+    if (hasLegacyConfig) {
+      const message = localize(
+        "ext.config.legacyPrettierConfigInUse",
+        LEGACY_VSCODE_PRETTIER_CONFIG_MESSAGE
+      );
+
+      const result = await window.showErrorMessage(
+        message,
+        MIGRATE_CONFIG_ACTION_TEXT
+      );
+      if (result && result === MIGRATE_CONFIG_ACTION_TEXT) {
+        await this.createConfigFileCommand(migratedOptions);
+        legacyConfigOptions.forEach(key => {
+          const inspected = vscodeConfig.inspect(key);
+          if (inspected && inspected.workspaceFolderValue) {
+            vscodeConfig.update(
+              key,
+              undefined,
+              ConfigurationTarget.WorkspaceFolder
+            );
+          }
+        });
+      }
+    }
+    return hasLegacyConfig;
   }
 }