+++ src/test/suite/notifications.test.ts
@@ -4,7 +4,7 @@ import * as sinon from "sinon";
 // tslint:disable-next-line: no-implicit-dependencies
 import { window } from "vscode";
 import {
-  LEGACY_VSCODE_CONFIG_MESSAGE,
+  LEGACY_VSCODE_PRETTIER_CONFIG_MESSAGE,
   OUTDATED_PRETTIER_VERSION_MESSAGE
 } from "../../Consts";
 import { format } from "./format.test";
@@ -24,7 +24,7 @@ suite("Test notifications", function() {
   });
   test("shows error for legacy vscode config", async () => {
     await format("outdated", "ugly.js");
-    assert(showInputBox.calledWith(LEGACY_VSCODE_CONFIG_MESSAGE));
+    assert(showInputBox.calledWith(LEGACY_VSCODE_PRETTIER_CONFIG_MESSAGE));
   });
   test("does not show error with valid project", async () => {
     await format("plugins", "index.php");